+++ backend/api/mapcreator/maps/__mocks__/maps.queries.js
@@ -1,0 +1,10 @@
+const mockConnectionQueries = {
+    insertMap: jest.fn(),
+    getMapInfo: jest.fn(),
+    getMapsByGameMapId: jest.fn(),
+    updateMapTitle: jest.fn().mockResolvedValue(true),
+    getAllImageIdsForMap: jest.fn(),
+    deleteMapById: jest.fn().mockResolvedValue(true)
+};
+
+module.exports = mockConnectionQueries;
