+++ backend/api/mapcreator/maps/maps.service.js
@@ -1,4 +1,5 @@
-const database = require("#sql/database.js");
+const database = require("#mapcreator/maps/maps.queries.js");
+const { getConnection, insertImage, deleteImageById, updateImagePath } = require("#sql/database.js");
 const AppError = require("#utils/app-error.js");
 const ERRORS = require("#utils/error-messages.js");
 const fs = require("fs/promises");
@@ -25,7 +26,7 @@ async function updateMap(userId, mapID, newTitle) {
             throw new AppError(ERRORS.MAP.NOT_FOUND, 404);
         }
 
-        dbConnection = await database.getConnection();
+        dbConnection = await getConnection();
         await dbConnection.beginTransaction();
 
         const updateSuccess = await database.updateMapTitle(dbConnection, mapID, newTitle);
@@ -63,10 +64,10 @@ async function createMap(userId, gameMapID, title, file) {
             throw new AppError(ERRORS.COMMON.IMAGE_PROCESSING_ERROR, 422);
         }
 
-        dbConnection = await database.getConnection();
+        dbConnection = await getConnection();
         await dbConnection.beginTransaction();
 
-        const imageId = await database.insertImage(dbConnection, imageData.width, imageData.height, "pending");
+        const imageId = await insertImage(dbConnection, imageData.width, imageData.height, "pending");
 
         const newMapId = await database.insertMap(dbConnection, title, gameMapID, imageId);
 
@@ -88,7 +89,7 @@ async function createMap(userId, gameMapID, title, file) {
         });
 
         let dbPath = path.join(relativeDestDir, processedImagePaths.targetFileName);
-        const updateSuccess = await database.updateImagePath(dbConnection, imageId, dbPath);
+        const updateSuccess = await updateImagePath(dbConnection, imageId, dbPath);
         if (!updateSuccess) {
             throw new AppError(ERRORS.MAP.SAVE_FAILED, 500);
         }
@@ -124,7 +125,7 @@ async function deleteMap(userId, mapID) {
             throw new AppError(ERRORS.MAP.NOT_FOUND, 404);
         }
 
-        dbConnection = await database.getConnection();
+        dbConnection = await getConnection();
         await dbConnection.beginTransaction();
 
         let imageIdsToDelete = await database.getAllImageIdsForMap(dbConnection, mapID);
@@ -135,7 +136,7 @@ async function deleteMap(userId, mapID) {
         }
 
         for (const imageId of imageIdsToDelete) {
-            let successImageDeletion = await database.deleteImageById(dbConnection, imageId);
+            let successImageDeletion = await deleteImageById(dbConnection, imageId);
             if (!successImageDeletion) {
                 throw new AppError(ERRORS.MAP.IMAGE_DELETIONS_FAILED, 500);
             }
