+++ backend/sql/__mocks__/database.js
@@ -13,16 +13,10 @@ const mockDatabase = {
     checkUserOwnsPoint: jest.fn().mockResolvedValue(true),
     checkUserOwnsConnection: jest.fn().mockResolvedValue(true),
 
-    updateMapTitle: jest.fn().mockResolvedValue(true),
     getGameMapDetails: jest.fn(),
-    getMapsByGameMapId: jest.fn(),
-    getMapInfo: jest.fn(),
-    insertMap: jest.fn(),
     insertImage: jest.fn(),
     updateImagePath: jest.fn().mockResolvedValue(true),
-    deleteMapById: jest.fn().mockResolvedValue(true),
     deleteImageById: jest.fn().mockResolvedValue(true),
-    getAllImageIdsForMap: jest.fn(),
     getPointsOnMap: jest.fn(),
     getGameMapIdByMapId: jest.fn(),
     getPointOnMapByCoordinates: jest.fn(),
